models/evaluation: scalability evaluation added; implementation and evaluation tests: small adaptions and additions and corrections

+++ tests/org.palladiosimulator.pcm.confidentiality.context.analysis.testframework/src/org/palladiosimulator/pcm/confidentiality/context/analysis/testframework/BaseTest.java
@@ -81,9 +81,11 @@ public abstract class BaseTest {
 
     @AfterEach
     protected void shutdownEval() {
-        this.eval.shutdown();
-        var file = new File(this.pathXACML);
-        file.delete();
+        if (this.eval != null) {
+            this.eval.shutdown();
+            var file = new File(this.pathXACML);
+            file.delete();
+        }
     }
